This update will change the image serving functionality
The update changes the image serving mechanism from QPixMap/QLabel to pyqtgraph/viewBox/numpy array. The reason for this is to be able to track cursor position
and serve image coordinates in realtime. The former solution was percieved to beslower than the pyqtgraph one. This change makes it easier to determine what values to chose for the spectral projection start an stopd.